+++ Changelog.md
@@ -1,4 +1,5 @@
 --------------------------Changelog for v1.9.4-beta1 of SecurityCraft--------------------------
 
+- New: Game rules "fakeWaterSourceConversion" and "fakeLavaSourceConversion" to control source conversion of fake liquids, akin to vanilla's water and lava
 - Change: Items in SecurityCraft's creative tabs have been reordered to match vanilla's order and in general make more sense
 - Change: Sounds for nether wood reinforced blocks have been adjusted to match the vanilla variants